initial commit
got somewhere with css properties- they don't exist unless the object is
linked on the DOM somewhere and "custom" properties don't seem to work
for chrome, so unused ones seem appropriate at the time being.

cannot sidestep local security issue with canvas and images... can
upload to *website* or run a python script that does a slow local server
to serve the page.

should be able to get somewhere with dynamically drawing digits in the
canvas tomorrow.