Start of server-side interaction
Want to try this with python+sqlite, so I can do it locally without
installing more software and offline.  Python's http server method is
more responsive and I don't need to wait the delay for a server to
update the most recent file.  Right now I have to do it on the live
server.  pwds.php is not included for obvious reasons.  Have to later
make sure this is XSS safe.