Tabbing in input of ":" character
Any type of invalid input in input.type=number will refuse to be pulled
from javascript (despite being shown in the input box), so the only way
to make "4:" work is to save the valid typed "4" and recover the value
when typing ":".  That is without resorting to using the generic text
input type and building everything around that.