Found a hacky way for CSS variables
CSS queries won't work unless there is an element the class belongs to.
And it looked like the last added class overrode all previous in the
case they specify the same property, so I couldn't add all classes to
one hidden element, but to an individual element for each colliding
property value.